added spellchecker, renamed variables
