+++ smells-like-lucene-it/src/main/java/tcd/parse/ReplaceTerm.java
@@ -1,4 +1,4 @@
-package tcd.parse_to_index;
+package tcd.parse;
 
 public class ReplaceTerm {
 	private String replaceFrom;
